Issue: Non-required fields were not serializing properly.
Explanation:
Suppose an entity has a non required field.
When the default schema serializer tries to serialize an entity from a dictionary, it expects the dictionary to have a key for the non required field.
This is not the expected behavior.
The entity should get serialized even if the dictionary is missing the keys for non required fields.
It should throw an error only if the field is a required field.

Solution:
Check whether a field is required, before throwing exceptions if the key is missing.

Code changes.
Catch exceptions thrown by the attribute getter, and bubble them up only if the field is required.
Suppress errors if field is not a required field.
rip/generic_steps/default_schema_serializer.py

Unit tests.
tests/unit_tests/generic_steps/test_default_schema_serializer.py